trivia-agent/requirements.txt
Flask
google-cloud-aiplatform
Flask-CORS
